# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/clear": {
            "get": {
                "description": "Clear all database",
                "tags": [
                    "debug"
                ],
                "summary": "Clear all database",
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "description": "Allows user to log into the system",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Login",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User login",
                        "name": "username",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "User password",
                        "name": "password",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "description": "Allows user to log out from the system",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Logout",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/private/deleteme": {
            "post": {
                "description": "Removes user from system",
                "tags": [
                    "user"
                ],
                "summary": "Delete user",
                "responses": {}
            }
        },
        "/api/private/me": {
            "post": {
                "description": "Allows user to modify its data",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Modify user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User login",
                        "name": "username",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "User name",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "User password",
                        "name": "password",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "User email",
                        "name": "email",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "Show name or not",
                        "name": "showName",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/register": {
            "post": {
                "description": "Add new user to system",
                "tags": [
                    "user"
                ],
                "summary": "Register new user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User login",
                        "name": "username",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Email address",
                        "name": "email",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "User password",
                        "name": "password",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "406": {
                        "description": "Not Acceptable",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/articles/": {
            "get": {
                "description": "Returns all articles in system",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "articles"
                ],
                "summary": "Returns all articles in system",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.ArticleResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Allows to add article to system",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "articles"
                ],
                "summary": "Add article",
                "parameters": [
                    {
                        "description": "Article to add",
                        "name": "article",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/server.ArticleJsonBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/articles/:articleId": {
            "get": {
                "description": "Returns full article data with array of its comments",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "articles"
                ],
                "summary": "Returns article with comments",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Article ID",
                        "name": "articleId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.FullEntryResponseSwagger"
                        }
                    }
                }
            }
        },
        "/articles/page/:page": {
            "get": {
                "description": "Returns articles with pagination",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "articles"
                ],
                "summary": "Returns articles with pagination",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page to display",
                        "name": "page",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.ArticleResponse"
                        }
                    }
                }
            }
        },
        "/comment": {
            "get": {
                "description": "Get call comments in system",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "articles"
                ],
                "summary": "Get comments",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.CommentResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Add comment for article",
                "tags": [
                    "articles"
                ],
                "summary": "Add comment",
                "parameters": [
                    {
                        "description": "Comment",
                        "name": "comment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/server.CommentJsonBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/comment/get/:articleId": {
            "get": {
                "description": "Returns all comments for article",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "articles"
                ],
                "summary": "Returns article comments",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Article ID",
                        "name": "articleId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.CommentResponse"
                        }
                    }
                }
            }
        },
        "/comment/like/:commentId": {
            "get": {
                "description": "Add dislike to comment",
                "tags": [
                    "articles"
                ],
                "summary": "Add dislike",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Comment ID",
                        "name": "commentId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        }
    },
    "definitions": {
        "server.ArticleJsonBody": {
            "type": "object",
            "properties": {
                "text": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "server.ArticleResponse": {
            "type": "object",
            "properties": {
                "AddDate": {
                    "type": "string"
                },
                "AuthorId": {
                    "type": "string"
                },
                "AuthorName": {
                    "type": "string"
                },
                "ID": {
                    "type": "integer"
                },
                "Text": {
                    "type": "string"
                },
                "Title": {
                    "type": "string"
                }
            }
        },
        "server.CommentJsonBody": {
            "type": "object",
            "properties": {
                "articleId": {
                    "type": "integer"
                },
                "authorName": {
                    "type": "string"
                },
                "text": {
                    "type": "string"
                }
            }
        },
        "server.CommentResponse": {
            "type": "object",
            "properties": {
                "ArticleId": {
                    "type": "integer"
                },
                "AuthorName": {
                    "type": "string"
                },
                "Dislikes": {
                    "type": "integer"
                },
                "ID": {
                    "type": "integer"
                },
                "Likes": {
                    "type": "integer"
                },
                "Text": {
                    "type": "string"
                }
            }
        },
        "server.FullEntryResponseSwagger": {
            "type": "object",
            "properties": {
                "article": {
                    "$ref": "#/definitions/server.ArticleResponse"
                },
                "comments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/server.CommentResponse"
                    }
                }
            }
        }
    }
}